==> code/07-Dataset-construction-summary.qmd <==
---
title: "Dataset construction summary"
format: html
---

## Purpose

Explain how I have created the new variables so be used. These are created with code below, stored separately from the original and then can be joined. Creating a codebook that can go along with the one from the Swedish census. 

Then this way it is easy to explain. I keep the variable names the same, and for the new ones that I create I keep a record of how they are created. 

This document then needs a summary with just the codebook.

```{r}
library(tidyverse)
df <- read.csv(here::here("data", "Folk1930.csv"), sep = "\t", fileEncoding="latin1")

df <- df %>% as_tibble() %>% 
  janitor::clean_names()
```


## Breakdown

### Age and adults

```{r}
codebook <- read_rds("data/codebook_1930.rds")

df_augmented <- df %>%
  mutate(
    yob = parse_number(fodar),
    age = 1930 - yob,
    adult_1930 = case_when(
      # turns 15 in 1930
      yob <= 1915 ~ "Adult",
      TRUE ~ "Youth"
    ),
    adult_1930 = factor(adult_1930)
  ) %>% 
  select(id, age, adult_1930)

# df_augmented %>% write_rds("data/augmented/age.rds")
```

## Lives in parish of birth

```{r}
df_augmented <- df %>%
  mutate(
    fscbkod = as.numeric(str_squish(fscbkod)),
    lives_in_parish_of_birth = case_when(
      scbkod != fscbkod ~ 0,
      scbkod == fscbkod ~ 1,
      TRUE ~ NA_real_
    )
  ) %>% 
  select(id, lives_in_parish_of_birth)

# df_augmented %>% write_rds("data/augmented/lives_in_parish_of_birth.rds")
```

## Creating indicator variables


```{r}
df_electricity_parishes <- read_rds("data/temp/electricity_parishes_1930.rds")

df_electricity_parishes <- df_electricity_parishes %>% 
  mutate(parish_code = as.numeric(str_remove(parish_code, "\\d\\d\\d$"))) %>% 
  mutate(electricity_parish = 1) %>% 
  distinct()

df_electricity_living_in <- df %>% 
  select(id, scbkod) %>% 
  left_join(df_electricity_parishes, by = c("scbkod" = "parish_code")) %>% 
  mutate(electricity_parish_living_in = case_when(
    electricity_parish == 1 ~ 1,
    TRUE ~ 0
  )) %>% 
  select(id, electricity_parish_living_in)

df_electricity_living_in %>% 
  count(electricity_parish_living_in)

df_electricity_born <- df %>% 
  select(id, fscbkod) %>% 
  mutate(fscbkod = as.numeric(fscbkod)) %>% 
  left_join(df_electricity_parishes, by = c("fscbkod" = "parish_code")) %>% 
  mutate(electricity_parish_born = case_when(
    electricity_parish == 1 ~ 1,
    TRUE ~ 0
  )) %>% 
  select(id, electricity_parish_born)

df_electricity_born %>% 
  count(electricity_parish_born)

df_augmented <- df_electricity_living_in %>% 
  left_join(df_electricity_born)


# df_augmented %>% write_rds("data/augmented/electricity_parish_classifications.rds")
```

How to create the matrix. 

Need three columns. All called group_something.

Contents are electricity parishes, exclusion zone, and control groups.


```{r}
parishes_1930 <- df %>% 
  # select parishes in the 1930 dataset to left join to later
  distinct(scbkod)

parish_buffer_matrix_long <- readRDS("~/Recon/SC-1930/Electricity-control-groups/parish_buffer_matrix_long.rds")

parish_buffer_matrix_long %>% 
  filter(type == "parishes_1930") %>% 
  select(parish_code, electricity_parish, buffer, value) %>%
  distinct() %>% 
  pivot_wider(names_from = buffer, values_from =value, names_prefix = "control_")

# so 47 are in the exclusion zone in 1930 census

parish_buffer_matrix_long %>%
  filter(type == "parishes_1930") %>%
  distinct(parish_code, exclusion_zone_indicator) %>%
  count(exclusion_zone_indicator)

exclusion_zone <- parish_buffer_matrix_long %>%
  filter(type == "parishes_1930") %>%
  distinct(parish_code, exclusion_zone_indicator) %>%
  filter(exclusion_zone_indicator == 1) %>%
  mutate(parish_code = as.numeric(str_remove(parish_code, "\\d\\d\\d$"))) %>%
  distinct(parish_code) %>% 
  mutate(group = "exclusion zone")

# there are 34 electricity parishes with 9 digit codes. let's trim

eps <- parish_buffer_matrix_long %>%
  filter(type == "parishes_1930") %>%
  distinct(parish_code, electricity_parish) %>% 
  filter(electricity_parish == 1) %>% 
  mutate(parish_code = as.numeric(str_remove(parish_code, "\\d\\d\\d$"))) %>% 
  distinct(parish_code) %>% 
  mutate(group = "treatment")

# now the actual control groups

control_100 <- parish_buffer_matrix_long %>%
  filter(type == "parishes_1930") %>%
  mutate(parish_code = as.numeric(str_remove(parish_code, "\\d\\d\\d$"))) %>%
  distinct(parish_code, buffer) %>%
  filter(buffer == 100) %>%
  select(-buffer) %>%
  mutate(group = "control")

control_150 <- parish_buffer_matrix_long %>%
  filter(type == "parishes_1930") %>%
  mutate(parish_code = as.numeric(str_remove(parish_code, "\\d\\d\\d$"))) %>%
  distinct(parish_code, buffer) %>%
  filter(buffer == 150) %>%
  select(-buffer) %>%
  mutate(group = "control")

control_250 <- parish_buffer_matrix_long %>%
  filter(type == "parishes_1930") %>%
  mutate(parish_code = as.numeric(str_remove(parish_code, "\\d\\d\\d$"))) %>%
  distinct(parish_code, buffer) %>%
  filter(buffer == 250) %>%
  select(-buffer) %>%
  mutate(group = "control")
```

Now join them up without repeats.

188003

```{r}
# the problem is that there are some parishes that appear in multiple groups.
# I want to keep the treatment ones first, then then exclusion ones, then the control ones.
treat_group_100 <- eps %>%
  bind_rows(exclusion_zone) %>%
  distinct(parish_code, .keep_all = T) %>%
  bind_rows(control_100) %>%
  distinct(parish_code, .keep_all = T)


treat_group_100_augmented <- parishes_1930 %>% 
  left_join(treat_group_100, by = c("scbkod" = "parish_code"))

treat_group_150 <- eps %>%
  bind_rows(exclusion_zone) %>%
  distinct(parish_code, .keep_all = T) %>%
  bind_rows(control_150) %>%
  distinct(parish_code, .keep_all = T)

treat_group_150_augmented <- parishes_1930 %>% 
  left_join(treat_group_150, by = c("scbkod" = "parish_code"))

treat_group_250 <- eps %>%
  bind_rows(exclusion_zone) %>%
  distinct(parish_code, .keep_all = T) %>%
  bind_rows(control_250) %>%
  distinct(parish_code, .keep_all = T)

treat_group_250_augmented <- parishes_1930 %>% 
  left_join(treat_group_250, by = c("scbkod" = "parish_code"))
```

Now ready to save.

```{r}
treat_group_100_augmented <- df %>% 
  select(id, scbkod) %>% 
  left_join(treat_group_100, by = c("scbkod" = "parish_code")) %>% 
  rename(tce_group_100 = group)

treat_group_100_150_augmented <- treat_group_100_augmented %>% 
  inner_join(treat_group_150_augmented) %>% 
  rename(tce_group_150 = group)

treat_group_100_150_250_augmented <- treat_group_100_150_augmented %>% 
  inner_join(treat_group_250_augmented) %>% 
  rename(tce_group_250 = group) 

# treat_group_100_150_250_augmented %>% write_rds("data/augmented/treat_groups_100_150_250.rds")
```

### Now joining together the other variables

```{r}
df <- read.csv(here::here("data", "Folk1930.csv"), sep = "\t", fileEncoding="latin1")

df <- df %>% as_tibble() %>% 
  janitor::clean_names()

df_augmented <- read_rds("data/augmented/age.rds")

df <- df %>%
  left_join(df_augmented)

df_augmented <- read_rds("data/augmented/lives_in_parish_of_birth.rds")

df <- df %>%
  left_join(df_augmented)

df_augmented <- read_rds("data/augmented/electricity_parish_classifications.rds")

df <- df %>%
  left_join(df_augmented)

df_augmented <- read_rds("data/augmented/treat_groups_100_150_250.rds") %>% 
  select(-scbkod)

df <- df %>%
  left_join(df_augmented)

# df_augmented <- read_rds("data/augmented/hisco_5.rds")
# 
# df <- df %>%
#   left_join(df_augmented)

df_augmented <- read_rds("data/augmented/distance_from_birth_parish_to_current_parish.rds")

df_augmented <- df_augmented %>% 
  arrange(id, desc(dist_bp_to_cp_km)) %>% 
  distinct(id, .keep_all = T)

df <- df %>%
  left_join(df_augmented)


df_augmented <- read_rds("data/augmented/distance_from_previous_parish_to_current_parish.rds")

df_augmented <- df_augmented %>% 
  arrange(id, desc(dist_pp_to_cp_km)) %>% 
  distinct(id, .keep_all = T)

df <- df %>%
  left_join(df_augmented)

df_augmented <- read_rds(here::here("data", "occupations", "df_1930_assigned_occ_hisco_hisco_hisclass.rds"))

df <- df %>%
  left_join(df_augmented, by = c("id"))
```

This needs to be refactored by I can't do it right now.

```{r}
df <- df %>% 
  mutate(female = case_when(
    kon == "K" ~ 1,
    kon == "M" ~ 0,
    TRUE ~ NA_real_
  ))


# creating the hisclass groups
# following more power to the people
df <- df %>% 
  mutate(hisclass_code_abb = case_when(
    hisclass_code %in% c(1, 2) ~ 1,
    hisclass_code %in% c(3, 4, 5) ~ 2,
    hisclass_code == 6 ~ 3,
    hisclass_code == 7 ~ 4,
    hisclass_code == 8 ~ 5,
    hisclass_code %in% c(9, 10) ~ 6,
    hisclass_code %in% c(11, 12) ~ 7,
    TRUE ~ NA_real_
  ))

df <- df %>% 
  mutate(hisclass_group_abb = case_when(
    hisclass_code %in% c(1, 2) ~ "Elite",
    hisclass_code %in% c(3, 4, 5) ~ "White collar",
    hisclass_code == 6 ~ "Foremen",
    hisclass_code == 7 ~ "Medium-skilled workers",
    hisclass_code == 8 ~ "Farmers and fishermen",
    hisclass_code %in% c(9, 10) ~ "Low-skilled workers",
    hisclass_code %in% c(11, 12) ~ "Unskilled workers",
    TRUE ~ NA_character_
  ))

df <- df %>% 
  mutate(hisclass_group_abb = fct_reorder(hisclass_group_abb, hisclass_code_abb))

# code up schooling
df <- df %>%
  mutate(schooling = case_when(
    skola == "" ~ NA_character_,
    skola == "0" ~ "No school, no information about ability to read or write",
    skola == "1" ~ "No school, able to read and write",
    skola == "2" ~ "Can read but can’t write",
    skola == "3" ~ "Primary school",
    skola == "4" ~ "Additional course after primary school",
    skola == "5" ~ "Secondary school without high school exam",
    skola == "7|6" ~ "High school exam without exam from higher educational institution",
    skola == "8" ~ "Exam from higher educational institution, but not from a university",
    skola == "9" ~ "University",
    skola == "x|X" ~ "No school, no information about ability to read or write",
    skola == "y|Y" ~ "No school, able to read and write",
    TRUE ~ NA_character_
  )) %>%
  mutate(schooling_abb = case_when(
    schooling == "No school, no information about ability to read or write" ~ NA_character_,
    schooling %in% c("Can read but can’t write", "No school, able to read and write") ~ "Literate",
    schooling %in% c("Primary school") ~ "Primary school",
    schooling %in% c("Secondary school without high school exam", "Additional course after primary school") ~ "Post primary schooling",
    schooling %in% c("University", "Exam from higher educational institution, but not from a university") ~ "Post-secondary and university",
    TRUE ~ NA_character_
  ))
```

Write out

```{r}
library(haven)
df %>% 
  select(-yrke.x) %>% 
  rename(yrke = yrke.y) %>% 
  write_dta("data/stata/folk_1930_augmented.dta", version = 15)

df %>% write_rds("data/augmented/df_1930_augmented.rds", compress = "gz")
  
```



### Population for the other censuses

```{r}
df_pop <- read_rds("data/Population_by_parish_by_year.rds")
treat_group_100_150_250_augmented <- read_rds("data/augmented/treat_groups_100_150_250.rds")

treat_group_250_codes <- treat_group_100_150_250_augmented %>% 
  distinct(scbkod, tce_group_250, tce_group_100, tce_group_150)

df_pop <- df_pop %>% 
  janitor::clean_names() %>% 
  mutate(parse = parse_number(as.character(parse))) %>% 
  mutate(parish_code = as.numeric(str_remove(parse, "\\d\\d\\d$"))) %>% 
  group_by(year, parish_code) %>% 
  mutate(total_pop = sum(n)) %>% 
  ungroup() %>% 
  select(year, parish_code, total_pop) %>% 
  distinct() %>% 
  pivot_wider(names_from = year, names_prefix = "population_parish_", values_from = total_pop)

df_pop_comp <- df_pop %>% 
  inner_join(treat_group_250_codes, by = c("parish_code" = "scbkod"))


df %>% 
  select(scbkod, tce_group_250)

df_pop_comp %>%
  pivot_longer(contains("population_parish")) %>%
  mutate(year = as.numeric(str_extract(name, "\\d\\d\\d\\d"))) %>%
  filter(!is.na(tce_group_100)) %>%
  group_by(tce_group_100, year) %>%
  summarise(mean_pop = mean(value, na.rm = T)) %>%
  ungroup() %>%
  ggplot(aes(year, mean_pop, colour = tce_group_100)) +
  geom_point() +
  geom_line()

df_pop_comp %>%
  pivot_longer(contains("population_parish")) %>%
  mutate(year = as.numeric(str_extract(name, "\\d\\d\\d\\d"))) %>%
  filter(!is.na(tce_group_250)) %>%
  mutate(tce_group_250 = str_to_title(tce_group_250)) %>%
  ggplot(aes(year, value, colour = tce_group_250, group = parish_code)) +
  geom_point() +
  geom_line() +
  facet_wrap(~tce_group_250) +
  labs(
    x = NULL,
    y = "Population in each parish",
    colour = "Group"
  )

```
