==> code/08-census-codebook-prep.Rmd <==
---
title: "1930 Census Codebook"
author: "JJayes"
date: "2022-09-22"
output: html_document
---

## Purpose

Making a codebook of the 1930 census and explaining the new variables.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
library(tidyverse)
```


```{r}
df <- read_rds("data/augmented/df_1930_augmented.rds")

colnames_1930 <- colnames(df) %>%
  as_tibble() %>%
  rename(field = value)
```


```{r}
codebook <- readxl::read_excel("data/census_codebook.xlsx") %>% 
  janitor::clean_names()

codebook <- codebook %>% 
  mutate(field = janitor::make_clean_names(field))

# codebook %>% write_rds("data/codebook_1930.rds")
```


```{r}
codebook %>% 
  DT::datatable(rownames = F,
                colnames = c("Variable", "Data type", "Description", "Source"))
```


Create summary of each var

By numeric or factor. 

If numeric show five number summary. 

If factor, count factors and show top 3

```{r}
df %>% 
  select(where(is.numeric)) %>% 
  summarise()

df %>% 
  select(where(is.character)) %>% 
  summarise(n_distinct = n_distinct(arkbild))

```
