==> code/19-hisco-to-hisclass.qmd <==
---
title: "Untitled"
format: html
---

## Purpose

Translate HISCO to Hisclass.

## Context

Vinzent Ostermeyer has provided me with a do file that he created from the spss file from van Leeuwen and Maas.

The source of the van Leeuwen and Maas classification is at [this link](http://www.towardsopensocieties.org/Towards_Open_Societies/news/Entries/2011/4/22_New_HISCO_to_HISCLASS_file_files/hisco_hisclass12_book%40_numerical.inc) which no longer works.

The process of going from hisco to hisclass has three steps requiring three variables. I currently do not have the three variables needed, so I will proceed with just this one for now.

### Process

It's really just a lookup table. I don't know why it's formatted like this.

I'm just going to tidy it up and create a lookup table.

```{r}
text = c("(1110=2) (1120=2) (1130=2) (1140=2) (1150=2) (1190=2) (1210=2) ///
		(1220=2) (1230=2) (1240=2) (1250=2) (1260=2) (1270=2) (1280=2) ///
		(1290=2) (1320=2) (1330=2) (1340=2) (1350=2) (1390=2) (1400=4) ///
		(1420=4) (1430=4) (1490=4) (2000=2) (2120=2) (2130=2) (2140=2) ///
		(2210=2) (2220=2) (2230=2) (2235=2) (2240=2) (2245=2) (2250=2) ///
		(2255=2) (2260=4) (2290=2) (2305=2) (2310=2) (2320=2) (2330=2) ///
		(2340=2) (2390=2) (2410=2) (2420=2) (2430=2) (2440=2) (2450=2) ///
		(2460=2) (2470=2) (2480=2) (2485=2) (2490=2) (2510=2) (2520=2) ///
		(2590=2) (2620=2) (2630=2) (2690=2) (2710=2) (2720=2) (2730=2) ///
		(2740=2) (2790=2) (2810=2) (2820=2) (2830=2) (2890=2) (2920=2) ///
		(2930=2) (2940=2) (2950=2) (2990=2) (3010=4) (3020=4) (3030=4) ///
		(3040=4) (3050=2) (3090=4) (3110=4) (3120=4) (3130=4) (3140=4) ///
		(3150=4) (3160=4) (3170=4) (3180=4) (3190=4) (3200=4) (3310=4) ///
		(3320=4) (3330=4) (3340=4) (3390=4) (3405=4) (3410=4) (3420=4) ///
		(3430=4) (3490=4) (3510=4) (3520=4) (3530=4) (3540=4) (3550=4) ///
		(3590=4) (3610=4) (3620=4) (3690=4) (3700=4) (3720=4) (3730=4) ///
		(3790=4) (3810=4) (3820=4) (3890=4) (3920=4) (3930=4) (3990=4) ///
		(4120=2) (4125=2) (4130=2) (4140=2) (4150=2) (4160=2) (4200=3) ///
		(4215=3) (4217=3) (4220=3) (4230=3) (4240=4) (4250=3) (4290=3) ///
		(4315=1) (4320=2) (4330=2) (5110=2) (5120=2) (5130=2) (5190=2) ///
		(5220=2) (5230=2) (5240=2) (5250=2) (5260=2) (5265=2) (5270=2) ///
		(5280=2) (5290=2) (5320=2) (5330=2) (5340=2) (5350=2) (5360=4) ///
		(5390=2) (5400=4) (5420=4) (5430=4) (5490=4) (6100=2) (6105=2) ///
		(6110=2) (6120=2) (6130=2) (6140=2) (6190=2) (6210=4) (6310=2) ///
		(6320=2) (6330=2) (6420=4) (6430=4) (6440=4) (6510=2) (6520=2) ///
		(6590=2) (6610=4) (6710=2) (6810=4) (6910=4) (6920=2) (6990=2) ///
		(7110=4) (7120=4) (7130=4) (7140=4) (7190=4) (7210=9) (7310=4) ///
		(7410=9) (7520=2) (7530=4) (7620=4) (7630=4) (7640=9) (7710=4) ///
		(7920=4) (7930=4) (7940=4) (7950=4) (7990=4) (8110=2) (8120=2) ///
		(8130=2) (8220=2) (8230=2) (8240=2) (8250=2) (8310=2) (8420=2) ///
		(8490=2) (9010=2) (9020=2) (9030=2) (11010=2) (11020=2) (11090=2) ///
		(12000=2) (12110=2) (12210=2) (12310=2) (12410=2) (12910=2) (13000=2) ///
		(13020=2) (13030=2) (13100=2) (13120=2) (13125=2) (13130=2) (13140=2) ///
		(13145=2) (13150=2) (13155=2) (13160=2) (13165=2) (13170=2) (13175=2) ///
		(13180=2) (13190=2) (13200=2) (13215=2) (13220=2) (13230=2) (13240=2) ///
		(13250=2) (13260=2) (13270=2) (13275=2) (13280=2) (13290=2) (13320=4) ///
		(13390=4) (13420=4) (13490=5) (13520=4) (13530=4) (13540=4) (13590=4) ///
		(13920=2) (13930=2) (13940=1) (13950=2) (13960=2) (13990=2) (14120=2) ///
		(14130=4) (14140=4) (14190=2) (14920=5) (14990=2) (15120=2) (15130=2) ///
		(15900=2) (15915=2) (15920=2) (15925=4) (15930=2) (15935=2) (15940=2) ///
		(15945=2) (15950=2) (15955=2) (15960=2) (15965=2) (15990=2) (16000=4) ///
		(16100=4) (16120=4) (16130=4) (16140=4) (16150=4) (16160=4) (16190=4) ///
		(16200=4) (16220=4) (16230=2) (16240=4) (16250=4) (16290=4) (16310=4) ///
		(16320=4) (16330=4) (16340=4) (16350=4) (16360=4) (16370=4) (16390=4) ///
		(17000=4) (17120=2) (17130=2) (17135=3) (17140=4) (17145=4) (17150=4) ///
		(17190=2) (17220=2) (17230=4) (17320=4) (17330=1) (17340=3) (17350=4) ///
		(17390=4) (17420=4) (17430=4) (17440=2) (17490=4) (17520=4) (17530=4) ///
		(17540=7) (17550=7) (17590=7) (17920=5) (17930=5) (17940=5) (17990=4) ///
		(18020=9) (18030=4) (18040=4) (18050=4) (18090=5) (19120=4) (19130=4) ///
		(19140=2) (19190=4) (19220=2) (19230=2) (19240=2) (19250=2) (19260=2) ///
		(19270=2) (19290=2) (19310=4) (19320=4) (19330=4) (19340=4) (19350=4) ///
		(19390=4) (19420=4) (19430=2) (19440=2) (19490=4) (19520=2) (19530=4) ///
		(19540=2) (19920=4) (19930=4) (19940=3) (19950=2) (19960=5) (19990=4) ///
		(20110=1) (20210=1) (21000=3) (21110=1) (21200=3) (21220=3) (21230=3) ///
		(21240=3) (21300=3) (21320=1) (21330=3) (21340=3) (21420=1) (21430=3) ///
		(21440=3) (21490=3) (21920=1) (21940=1) (21950=1) (21960=1) (21970=1) ///
		(21980=1) (21990=1) (22000=3) (22110=3) (22190=3) (22210=3) (22220=3) ///
		(22230=3) (22235=4) (22240=4) (22245=4) (22250=3) (22255=3) (22260=3) ///
		(22265=4) (22270=4) (22275=3) (22280=3) (22290=3) (22320=3) (22330=3) ///
		(22390=3) (22420=6) (22425=6) (22430=6) (22440=6) (22450=6) (22460=6) ///
		(22490=6) (22520=3) (22610=3) (22620=6) (22630=6) (22640=6) (22650=6) ///
		(22655=6) (22660=6) (22665=6) (22670=6) (22675=6) (22680=6) (22690=6) ///
		(30000=5) (31000=5) (31020=4) (31030=2) (31040=4) (31090=4) (32110=5) ///
		(32120=5) (32130=5) (32140=5) (32150=5) (32190=5) (32220=5) (32290=5) ///
		(33110=4) (33120=4) (33130=5) (33135=5) (33140=5) (33150=5) (33160=5) ///
		(33170=4) (33190=5) (33920=5) (33930=5) (33940=5) (33990=5) (34120=5) ///
		(34130=5) (34220=5) (34230=5) (34290=5) (36000=9) (36020=9) (36030=9) ///
		(36040=9) (36090=9) (37000=9) (37010=4) (37020=9) (37030=9) (37040=9) ///
		(37090=9) (38000=4) (38020=9) (38030=4) (38040=5) (38050=4) (38060=4) ///
		(38090=9) (39110=5) (39120=5) (39130=5) (39140=5) (39150=9) (39190=5) ///
		(39220=5) (39230=5) (39310=5) (39320=5) (39330=5) (39340=4) (39350=5) ///
		(39390=5) (39410=5) (39420=5) (39430=5) (39440=5) (39490=5) (39520=5) ///
		(39530=5) (39920=5) (39930=5) (39940=5) (39950=9) (39960=5) (39990=5) ///
		(41020=3) (41025=4) (41030=4) (41040=4) (42220=4) (42230=4) (43120=4) ///
		(43130=4) (43200=4) (43220=4) (43230=4) (44000=4) (44100=4) (44120=4) ///
		(44130=4) (44140=4) (44190=4) (44220=4) (44230=4) (44320=4) (44330=2) ///
		(45120=4) (45125=5) (45130=5) (45140=5) (45150=5) (45190=5) (45220=11) ///
		(45230=11) (45240=11) (45290=11) (49020=5) (49030=11) (49090=5) (51000=4) ///
		(51020=1) (51030=4) (51040=4) (51050=4) (51090=4) (53100=7) (53120=6) ///
		(53130=7) (53140=7) (53150=9) (53190=11) (53210=9) (53220=3) (53230=9) ///
		(53240=9) (53250=9) (53290=9) (54010=9) (54020=11) (54030=9) (54035=9) ///
		(54040=5) (54050=11) (54055=9) (54060=5) (54090=9) (55100=5) (55120=5) ///
		(55125=5) (55130=5) (55140=5) (55190=5) (55220=11) (55230=11) (55240=11) ///
		(55290=11) (56010=9) (56020=9) (56030=9) (56040=9) (56050=9) (56060=9) ///
		(56065=9) (56070=9) (56080=9) (56090=11) (57020=9) (57025=9) (57030=9) ///
		(57040=9) (57050=9) (57060=9) (57070=9) (57090=7) (58110=7) (58120=4) ///
		(58130=7) (58140=7) (58220=4) (58225=4) (58230=4) (58240=5) (58300=9) ///
		(58320=1) (58330=3) (58340=9) (58920=4) (58930=5) (58940=5) (58990=5) ///
		(59100=5) (59120=5) (59130=5) (59220=4) (59230=7) (59290=9) (59920=5) ///
		(59930=5) (59940=9) (59950=9) (59960=9) (59970=9) (59980=11) (59990=9) ///
		(61110=8) (61115=10) (61120=1) (61220=8) (61230=8) (61235=8) (61240=8) ///
		(61250=8) (61260=8) (61270=8) (61290=8) (62105=12) (62110=12) (62120=10) ///
		(62210=12) (62220=12) (62230=12) (62240=12) (62250=10) (62260=12) (62290=12) ///
		(62320=10) (62330=12) (62340=12) (62350=12) (62390=12) (62400=12) (62410=12) ///
		(62420=12) (62430=12) (62440=12) (62450=12) (62460=12) (62490=12) (62510=10) ///
		(62520=10) (62590=12) (62610=12) (62620=12) (62690=12) (62700=12) (62720=12) ///
		(62730=12) (62740=12) (62790=12) (62820=10) (62890=10) (62920=10) (62930=12) ///
		(62940=12) (62950=12) (62960=11) (62970=11) (62990=12) (63110=10) (63120=10) ///
		(63130=8) (63140=10) (63190=10) (63220=3) (63230=12) (63240=8) (63250=12) ///
		(63290=12) (64100=10) (64120=10) (64130=10) (64920=12) (64930=12) (64940=10) ///
		(64945=10) (64950=10) (64960=10) (64990=8) (71105=9) (71110=9) (71120=9) ///
		(71125=9) (71130=9) (71140=9) (71150=7) (71160=9) (71170=11) (71190=11) ///
		(71200=9) (71220=9) (71230=9) (71240=9) (71250=9) (71260=9) (71270=9) ///
		(71280=9) (71290=11) (71300=9) (71320=9) (71330=9) (71340=7) (71350=7) ///
		(71360=9) (71370=9) (71380=9) (71390=9) (72000=9) (72100=9) (72120=9) ///
		(72130=9) (72140=9) (72150=9) (72160=9) (72170=9) (72190=9) (72200=9) ///
		(72220=9) (72230=9) (72240=9) (72250=9) (72260=9) (72270=9) (72290=9) ///
		(72320=9) (72330=9) (72340=9) (72390=9) (72420=9) (72430=9) (72440=9) ///
		(72450=9) (72490=9) (72500=7) (72520=7) (72530=7) (72540=9) (72550=7) ///
		(72555=7) (72560=9) (72590=9) (72620=9) (72630=9) (72640=9) (72650=9) ///
		(72690=11) (72720=9) (72725=9) (72730=9) (72740=9) (72750=9) (72790=11) ///
		(72820=9) (72830=9) (72840=9) (72850=9) (72890=9) (72920=9) (72930=11) ///
		(72940=9) (72990=9) (73100=9) (73120=9) (73130=9) (73190=11) (73200=9) ///
		(73210=9) (73220=9) (73230=9) (73240=9) (73250=11) (73260=9) (73270=9) ///
		(73290=9) (73320=9) (73330=9) (73340=9) (73350=9) (73360=9) (73390=11) ///
		(73400=9) (73420=9) (73430=9) (73440=9) (73450=9) (73460=9) (73490=9) ///
		(74100=9) (74120=9) (74130=9) (74140=9) (74190=9) (74220=9) (74230=9) ///
		(74240=9) (74290=9) (74320=9) (74330=9) (74340=9) (74350=9) (74390=9) ///
		(74400=9) (74420=9) (74430=9) (74440=9) (74450=9) (74460=9) (74470=9) ///
		(74490=9) (74500=11) (74520=9) (74530=9) (74540=9) (74550=9) (74560=9) ///
		(74570=9) (74590=9) (74915=9) (74920=9) (74925=9) (74930=9) (74935=9) ///
		(74940=7) (74990=9) (75000=9) (75100=9) (75115=9) (75120=9) (75125=9) ///
		(75130=9) (75135=9) (75140=9) (75145=9) (75150=9) (75155=9) (75190=11) ///
		(75220=9) (75230=9) (75240=9) (75250=9) (75290=11) (75320=7) (75330=7) ///
		(75340=9) (75350=9) (75390=9) (75400=9) (75415=9) (75420=11) (75422=11) ///
		(75425=11) (75430=9) (75432=9) (75435=9) (75440=9) (75445=9) (75450=9) ///
		(75452=9) (75455=9) (75457=9) (75460=9) (75465=9) (75470=9) (75475=7) ///
		(75490=9) (75500=9) (75520=9) (75530=9) (75535=9) (75540=9) (75550=9) ///
		(75590=9) (75600=9) (75615=9) (75620=9) (75622=9) (75625=9) (75630=9) ///
		(75635=9) (75640=9) (75645=9) (75650=9) (75655=9) (75660=9) (75665=9) ///
		(75670=9) (75690=9) (75710=9) (75720=9) (75920=11) (75922=11) (75925=7) ///
		(75930=9) (75935=9) (75940=9) (75945=9) (75950=9) (75990=9) (76100=9) ///
		(76120=9) (76125=9) (76130=9) (76135=9) (76140=11) (76145=9) (76150=11) ///
		(76155=9) (76190=9) (76200=9) (76220=9) (76230=9) (76240=9) (76250=9) ///
		(76260=9) (76290=9) (77120=7) (77130=7) (77140=9) (77190=9) (77200=9) ///
		(77220=9) (77230=9) (77240=9) (77250=9) (77260=9) (77290=9) (77310=7) ///
		(77320=9) (77330=9) (77340=9) (77390=9) (77410=9) (77415=9) (77420=9) ///
		(77430=11) (77440=9) (77450=9) (77460=9) (77490=11) (77510=9) (77520=9) ///
		(77530=9) (77540=7) (77550=7) (77590=11) (77610=7) (77620=7) (77630=7) ///
		(77640=9) (77650=9) (77660=9) (77690=9) (77720=4) (77730=9) (77740=9) ///
		(77750=9) (77760=9) (77790=9) (77810=7) (77820=9) (77825=9) (77830=9) ///
		(77835=9) (77840=9) (77850=9) (77860=4) (77870=9) (77880=9) (77890=9) ///
		(77920=9) (77925=9) (77930=9) (77940=9) (77990=11) (78100=9) (78120=9) ///
		(78130=7) (78140=9) (78150=11) (78155=11) (78160=9) (78170=9) (78190=9) ///
		(78200=9) (78220=9) (78230=9) (78290=9) (78320=9) (78390=9) (78920=9) ///
		(78990=9) (79100=7) (79120=7) (79130=7) (79140=7) (79190=7) (79200=9) ///
		(79220=7) (79230=7) (79240=9) (79250=7) (79260=9) (79290=9) (79310=7) ///
		(79320=7) (79330=9) (79390=9) (79400=9) (79420=9) (79430=9) (79440=9) ///
		(79450=9) (79460=9) (79470=9) (79475=9) (79480=9) (79490=9) (79500=9) ///
		(79510=9) (79520=9) (79530=9) (79540=7) (79550=9) (79560=7) (79565=7) ///
		(79570=9) (79590=9) (79600=7) (79620=7) (79630=7) (79640=9) (79690=11) ///
		(79920=7) (79930=7) (79990=9) (80110=7) (80120=7) (80130=7) (80190=7) ///
		(80200=9) (80215=9) (80220=9) (80222=9) (80225=9) (80230=9) (80235=9) ///
		(80240=9) (80245=9) (80250=9) (80255=9) (80290=9) (80310=9) (80320=7) ///
		(80330=9) (80340=9) (80350=9) (80360=9) (80390=9) (81000=9) (81120=7) ///
		(81190=7) (81205=7) (81208=9) (81210=9) (81220=11) (81225=11) (81230=9) ///
		(81240=9) (81250=9) (81260=9) (81270=9) (81275=9) (81280=7) (81290=11) ///
		(81920=7) (81925=7) (81930=7) (81935=7) (81940=7) (81945=7) (81950=9) ///
		(81955=7) (81960=7) (81965=7) (81990=9) (82000=9) (82020=9) (82030=9) ///
		(82040=7) (82050=9) (82060=7) (82070=7) (82080=7) (82090=9) (83000=9) ///
		(83110=7) (83120=7) (83130=9) (83140=9) (83190=9) (83210=7) (83220=7) ///
		(83230=7) (83240=7) (83250=7) (83290=7) (83305=7) (83310=7) (83320=7) ///
		(83330=9) (83340=9) (83350=7) (83360=9) (83370=7) (83380=9) (83385=9) ///
		(83390=9) (83410=7) (83420=9) (83430=9) (83440=9) (83450=9) (83460=9) ///
		(83465=7) (83470=9) (83475=9) (83480=9) (83490=9) (83520=9) (83530=9) ///
		(83540=9) (83550=9) (83560=7) (83590=9) (83915=7) (83920=7) (83930=7) ///
		(83940=9) (83950=7) (83960=9) (83970=9) (83980=9) (83990=7) (84100=7) ///
		(84105=7) (84110=7) (84115=7) (84120=7) (84125=7) (84130=7) (84135=7) ///
		(84140=7) (84145=7) (84150=7) (84155=7) (84160=7) (84165=7) (84170=9) ///
		(84175=7) (84180=7) (84185=7) (84190=9) (84220=7) (84222=7) (84225=7) ///
		(84230=7) (84235=7) (84240=7) (84245=7) (84250=7) (84290=7) (84300=7) ///
		(84320=7) (84330=7) (84340=7) (84390=9) (84400=7) (84410=7) (84420=7) ///
		(84490=7) (84900=7) (84910=7) (84915=7) (84920=7) (84925=7) (84930=7) ///
		(84935=7) (84940=7) (84945=7) (84950=7) (84955=7) (84960=7) (84965=7) ///
		(84970=7) (84975=9) (84980=9) (84985=7) (84990=7) (85110=7) (85120=7) ///
		(85130=7) (85140=9) (85150=7) (85160=7) (85190=7) (85210=7) (85220=7) ///
		(85230=7) (85240=7) (85250=7) (85260=7) (85290=7) (85320=9) (85330=9) ///
		(85340=9) (85345=9) (85350=9) (85390=9) (85420=7) (85490=7) (85510=7) ///
		(85520=7) (85530=7) (85535=7) (85540=7) (85550=7) (85560=7) (85570=7) ///
		(85590=7) (85600=7) (85620=7) (85630=7) (85690=7) (85700=7) (85720=7) ///
		(85730=7) (85740=7) (85750=7) (85790=7) (85920=7) (85990=7) (86120=7) ///
		(86130=7) (86190=7) (86220=5) (86230=5) (86240=5) (86290=5) (87105=7) ///
		(87110=7) (87120=7) (87130=7) (87140=7) (87190=7) (87210=9) (87215=9) ///
		(87220=9) (87222=9) (87225=9) (87230=9) (87235=9) (87240=7) (87245=9) ///
		(87250=9) (87255=9) (87260=7) (87290=9) (87310=7) (87320=7) (87330=7) ///
		(87340=7) (87350=7) (87360=7) (87370=7) (87380=7) (87390=7) (87400=9) ///
		(87420=7) (87430=7) (87440=9) (87450=7) (87455=7) (87460=9) (87462=9) ///
		(87465=9) (87470=9) (87490=9) (88010=7) (88020=7) (88030=7) (88040=7) ///
		(88050=7) (88060=9) (88070=7) (88080=7) (88090=7) (89100=9) (89120=7) ///
		(89124=7) (89128=9) (89132=9) (89136=9) (89140=9) (89144=9) (89148=9) ///
		(89152=9) (89156=9) (89160=9) (89164=9) (89168=7) (89172=9) (89176=9) ///
		(89180=9) (89190=9) (89200=9) (89210=7) (89215=7) (89220=7) (89225=9) ///
		(89230=9) (89235=9) (89240=9) (89242=9) (89245=7) (89247=7) (89250=9) ///
		(89255=9) (89260=9) (89290=9) (89300=9) (89320=9) (89330=9) (89340=9) ///
		(89350=9) (89360=9) (89390=11) (89400=7) (89420=7) (89430=7) (89440=7) ///
		(89490=9) (89500=7) (89520=7) (89530=7) (89540=7) (89550=9) (89560=9) ///
		(89570=9) (89590=11) (89920=9) (89930=9) (89940=9) (89950=9) (89960=9) ///
		(89970=9) (89990=9) (90100=9) (90120=9) (90125=9) (90130=9) (90135=9) ///
		(90140=9) (90150=9) (90155=9) (90160=9) (90165=9) (90170=9) (90180=7) ///
		(90190=9) (90200=9) (90220=9) (90230=9) (90240=9) (90290=11) (91020=9) ///
		(91025=9) (91030=9) (91040=9) (91050=9) (91060=9) (91070=9) (91075=9) ///
		(91080=9) (91090=9) (92110=7) (92120=7) (92130=9) (92135=9) (92140=9) ///
		(92145=7) (92150=7) (92155=9) (92190=7) (92220=9) (92225=9) (92230=9) ///
		(92240=9) (92250=9) (92260=9) (92270=9) (92290=9) (92320=7) (92330=7) ///
		(92390=7) (92400=9) (92415=7) (92420=7) (92430=7) (92440=9) (92445=9) ///
		(92450=7) (92460=9) (92490=9) (92510=7) (92520=4) (92530=7) (92540=7) ///
		(92550=7) (92560=7) (92590=7) (92620=7) (92625=7) (92630=9) (92640=9) ///
		(92645=9) (92650=9) (92690=7) (92720=9) (92730=7) (92740=9) (92750=7) ///
		(92790=9) (92920=7) (92930=7) (92940=7) (92950=7) (92990=7) (93120=9) ///
		(93130=9) (93190=7) (93920=7) (93930=9) (93940=9) (93950=7) (93960=7) ///
		(93990=11) (94100=7) (94120=7) (94130=7) (94140=7) (94150=7) (94160=7) ///
		(94170=7) (94180=7) (94190=7) (94220=9) (94230=9) (94240=9) (94250=9) ///
		(94290=9) (94320=9) (94330=9) (94340=9) (94390=9) (94900=9) (94920=7) ///
		(94930=11) (94940=9) (94950=7) (94960=9) (94970=9) (94980=9) (94990=11) ///
		(95000=9) (95120=7) (95125=7) (95130=7) (95135=7) (95140=7) (95145=7) ///
		(95150=7) (95155=7) (95160=9) (95190=7) (95210=7) (95220=7) (95230=9) ///
		(95240=9) (95250=7) (95290=9) (95300=9) (95320=9) (95330=9) (95340=9) ///
		(95350=9) (95360=9) (95390=9) (95410=7) (95415=7) (95420=7) (95430=7) ///
		(95440=7) (95445=7) (95450=7) (95455=7) (95460=7) (95470=7) (95475=7) ///
		(95490=7) (95510=9) (95520=9) (95530=9) (95540=9) (95590=9) (95620=9) ///
		(95630=9) (95640=7) (95650=7) (95660=9) (95690=7) (95720=7) (95730=7) ///
		(95740=7) (95750=7) (95760=9) (95790=9) (95910=9) (95920=9) (95925=9) ///
		(95930=9) (95935=9) (95940=9) (95945=9) (95950=9) (95955=11) (95960=7) ///
		(95970=7) (95975=9) (95990=9) (96100=7) (96120=7) (96130=7) (96140=7) ///
		(96150=7) (96160=7) (96170=4) (96190=7) (96910=9) (96920=9) (96925=9) ///
		(96930=9) (96940=9) (96950=9) (96960=9) (96970=9) (96980=9) (96990=9) ///
		(97120=11) (97125=11) (97130=11) (97135=11) (97140=9) (97145=11) (97150=11) ///
		(97152=11) (97155=9) (97160=9) (97170=9) (97190=11) (97205=7) (97210=7) ///
		(97220=7) (97230=7) (97240=7) (97250=9) (97260=7) (97290=11) (97315=9) ///
		(97320=9) (97325=9) (97327=9) (97330=9) (97335=9) (97340=9) (97345=11) ///
		(97350=11) (97355=9) (97390=9) (97400=9) (97415=11) (97420=9) (97425=9) ///
		(97430=9) (97435=9) (97440=9) (97445=9) (97450=9) (97455=9) (97460=9) ///
		(97470=9) (97475=9) (97490=9) (97900=9) (97920=9) (97930=9) (97940=9) ///
		(97950=9) (97990=9) (98000=9) (98100=11) (98120=6) (98130=9) (98135=11) ///
		(98140=11) (98190=11) (98220=9) (98230=9) (98290=11) (98300=7) (98320=7) ///
		(98330=7) (98340=7) (98350=9) (98360=7) (98390=9) (98420=9) (98430=9) ///
		(98440=9) (98450=9) (98500=9) (98520=9) (98530=9) (98540=9) (98550=9) ///
		(98555=9) (98560=9) (98570=9) (98590=9) (98600=9) (98620=9) (98630=9) ///
		(98690=9) (98920=3) (98930=5) (98940=5) (98950=9) (98990=11) (99900=11) ///
		(99910=11) (99920=11) (99930=11)")
```


```{r}
hisco_hisclass_crosswalk <- text %>% 
  as_tibble() %>% 
  separate_rows(value, sep = "///") %>% 
  mutate(value = str_squish(value)) %>% 
  separate_rows(value, sep = " ") %>% 
  mutate(value = str_remove_all(value, "\\(|\\)")) %>% 
  # silly! must specify the break
  separate(value, into = c("hisco", "hisclass"), sep = "=") %>% 
  mutate(across(c(hisco, hisclass), as.numeric))
```

```{r}
# hisco_hisclass_crosswalk %>% write_rds("data/occupations/hisco_hisclass_crosswalk.rds")
```

```{r}
hisco_hisclass_crosswalk
```
